combine movies and user classes
