+++ app/components/MovieDetail/MovieDetail.js
@@ -1,0 +1,10 @@
+import React, { Component } from 'react';
+
+export default class MovieDetail extends Component {
+  constructor() {
+    super();
+    this.state {
+      detail: 
+    }
+  }
+}
